requirements.txt
openai
PyYAML
Jinja2
json-repair
mcp
requests
typer
rich
prompt_toolkit
